requirements.txt
aiogram>=3.2.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0 